+++ deploy/helpers/verifiers/passport.ts
@@ -46,7 +46,13 @@ import {
   PPerPassport_20_256_3_5_336_72_NAVerifier2__factory,
   PPerPassport_21_256_3_5_576_232_NAVerifier2__factory,
   PPerPassport_4_160_3_3_336_216_1_1296_3_256Verifier2__factory,
-  NoirRegisterIdentity_2_256_3_6_264_336_20_2448_5_2008__factory,
+  NoirRegisterIdentity_2_256_3_6_248_336_1_2432_3_256__factory,
+  NoirRegisterIdentity_2_256_3_6_336_264_21_2448_6_2008__factory,
+  NoirRegisterIdentity_2_256_3_6_336_248_1_2432_3_256__factory,
+  NoirRegisterIdentity_20_256_3_3_336_224_NA__factory,
+  NoirRegisterIdentity_1_256_3_4_600_248_1_1496_3_256__factory,
+  NoirRegisterIdentity_10_256_3_3_576_248_1_1184_5_264__factory,
+  NoirRegisterIdentity_21_256_3_3_576_232_NA__factory,
 } from "@ethers-v6";
 
 export const deployPVerifiers = async (deployer: Deployer) => {
@@ -100,5 +106,11 @@ export const deployPVerifiers = async (deployer: Deployer) => {
   await deployer.deploy(PMNEOptVerifier2__factory);
   await deployer.deploy(PMNEOpt2Verifier2__factory);
 
-  await deployer.deploy(NoirRegisterIdentity_2_256_3_6_264_336_20_2448_5_2008__factory);
+  await deployer.deploy(NoirRegisterIdentity_1_256_3_4_600_248_1_1496_3_256__factory);
+  await deployer.deploy(NoirRegisterIdentity_2_256_3_6_248_336_1_2432_3_256__factory);
+  await deployer.deploy(NoirRegisterIdentity_2_256_3_6_336_248_1_2432_3_256__factory);
+  await deployer.deploy(NoirRegisterIdentity_2_256_3_6_336_264_21_2448_6_2008__factory);
+  await deployer.deploy(NoirRegisterIdentity_10_256_3_3_576_248_1_1184_5_264__factory);
+  await deployer.deploy(NoirRegisterIdentity_20_256_3_3_336_224_NA__factory);
+  await deployer.deploy(NoirRegisterIdentity_21_256_3_3_576_232_NA__factory);
 };